правки з логуванням + фронт
