+++ App/Models/Model.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Models;
+
+class Model{
+
+    protected $pdo;
+
+    public function __construct($pdo) {
+        $this->pdo = $pdo;
+    }
+
+}
