+++ App/Models/Product.php
@@ -2,13 +2,11 @@
 
 namespace App\Models;
 use PDOException;
-
-class Product{
-
-    private $pdo;
+use Log;
+class Product extends Model{
 
     public function __construct($pdo) {
-        $this->pdo = $pdo;
+        parent::__construct($pdo);
     }
 
     public function getProducsWithCategories(array $get):array{      
@@ -56,29 +54,10 @@ class Product{
             $data = $request->fetchAll();
             return ['success' => true, 'data' => $data];
         } catch (PDOException $e) {
-            //TODO записати це в логи
-            echo 'Помилка підключення до бази даних: ' . $e->getMessage();
+            $logger = new Log();
+            $logger->filelog('Помилка підключення до бази даних: ' . $e->getMessage());
             return ['success' => false, 'data' => []]; 
         }
     }
 
-
-    public function getCategory() {        
-        $query = $this->pdo->query('SELECT categories.name AS category_name, 
-            categories.id AS category_id,  
-            COUNT(products.id) AS product_count
-            FROM categories
-            LEFT JOIN products ON categories.id = products.category_id
-            GROUP BY categories.id, categories.name;'); 
-
-        try {
-            $data = $query->fetchAll();
-            return ['success' => true, 'data' => $data];
-        } catch (PDOException $e) {
-            //TODO записати це в логи
-            echo 'Помилка підключення до бази даних: ' . $e->getMessage();
-            return ['success' => false, 'data' => []]; 
-        }
-
-    }
 }
